backend/flask/src/requirements.txt
flask
flask-restful
gunicorn
psycopg2-binary
python-dotenv
sqlalchemy